feat(ex/phoenix): Convert ErrorHelpers#error_tag into a component
