+++ lib/skate_web.ex
@@ -43,7 +43,6 @@ defmodule SkateWeb do
         only: [get_csrf_token: 0, view_module: 1, view_template: 1]
 
       import SkateWeb.CoreComponents
-      import SkateWeb.ErrorHelpers
 
       unquote(verified_routes())
     end
